+++ src/test/scala/au/com/agl/arc/load/HTTPLoadSuite.scala
@@ -123,7 +123,7 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
         inputView=outputView, 
         outputURI=new URI(s"${uri}/success/"), // ensure trailing slash to avoid 302 redirect
         headers=Map.empty,
-        validStatusCodes=None,
+        validStatusCodes=200 :: 201 :: 202 :: Nil,
         params=Map.empty
       )
     )
@@ -145,12 +145,12 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
           inputView=outputView, 
           outputURI=new URI(s"${uri}/failure/"), // ensure trailing slash to avoid 302 redirect
           headers=Map.empty,
-          validStatusCodes=None,
+          validStatusCodes=200 :: 201 :: 202 :: Nil,
           params=Map.empty
         )
-      )
+      ).get.count
     }
-    assert(thrown.getMessage == "HTTPLoad expects all response StatusCode(s) in [200, 201, 202] but server responded with [1 reponses 200 (OK), 1 reponses 401 (Unauthorized)].")        
+    assert(thrown.getMessage.contains("HTTPLoad expects all response StatusCode(s) in [200, 201, 202] but server responded with 401 (Unauthorized)."))
   }
 
   test("HTTPLoad: validStatusCodes") {
@@ -168,7 +168,7 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
         inputView=outputView, 
         outputURI=new URI(s"${uri}/failure/"), // ensure trailing slash to avoid 302 redirect
         headers=Map(key -> value),
-        validStatusCodes=Option(List(200,401)),
+        validStatusCodes=200 :: 401 :: Nil,
         params=Map.empty
       )
     )
@@ -189,7 +189,7 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
         inputView=outputView, 
         outputURI=new URI(s"${uri}/headers/"), // ensure trailing slash to avoid 302 redirect
         headers=Map(key -> value),
-        validStatusCodes=Option(List(200)),
+        validStatusCodes=200 :: Nil,
         params=Map.empty
       )
     )
@@ -211,12 +211,12 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
           inputView=outputView, 
           outputURI=new URI(s"${uri}/headers/"), // ensure trailing slash to avoid 302 redirect
           headers=Map(key -> "wrong"),
-          validStatusCodes=Option(List(200)),
+          validStatusCodes=200 :: Nil,
           params=Map.empty
         )
-      )
+      ).get.count
     }
-    assert(thrown.getMessage == "HTTPLoad expects all response StatusCode(s) in [200] but server responded with [2 reponses 401 (Unauthorized)].")            
+    assert(thrown.getMessage.contains("HTTPLoad expects all response StatusCode(s) in [200] but server responded with 401 (Unauthorized)."))      
   } 
 
   test("HTTPLoad: invalid inputView") {
@@ -235,7 +235,7 @@ class HTTPLoadSuite extends FunSuite with BeforeAndAfter {
           inputView=outputView, 
           outputURI=new URI(s"${uri}/success/"), // ensure trailing slash to avoid 302 redirect
           headers=Map.empty,
-          validStatusCodes=None,
+          validStatusCodes=200 :: 201 :: 202 :: Nil,
           params=Map.empty
         )
       )