+++ src/test/scala/au/com/agl/arc/transform/HTTPTransformSuite.scala
@@ -55,6 +55,7 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
   val binary = "binary"
   val empty = "empty"
   val body = "testpayload"
+  val delimiter = "\n"
 
   before {
     implicit val spark = SparkSession
@@ -107,9 +108,10 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
     dataset.createOrReplaceTempView(inputView)
     var payloadDataset = spark.sql(s"""
       SELECT *, TO_JSON(NAMED_STRUCT('dateDatum', dateDatum)) AS value FROM ${inputView}
-    """)
+    """).repartition(1)
     val inputDataset = MetadataUtils.setMetadata(payloadDataset, Extract.toStructType(cols.right.getOrElse(Nil)))
     inputDataset.createOrReplaceTempView(inputView)
+
 
     val transformDataset = transform.HTTPTransform.transform(
       HTTPTransform(
@@ -122,9 +124,13 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
         outputView=outputView,
         params=Map.empty,
         persist=false,
-        inputField="value"
+        inputField="value",
+        batchSize=2,
+        delimiter=delimiter
       )
     ).get
+
+    transformDataset.cache.count
 
     val expected = transformDataset.select(col("value"))
     val actual = transformDataset.select(col("body")).withColumnRenamed("body", "value")
@@ -148,7 +154,7 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
     dataset.createOrReplaceTempView(inputView)
     var payloadDataset = spark.sql(s"""
       SELECT *, TO_JSON(NAMED_STRUCT('dateDatum', dateDatum)) AS inputField FROM ${inputView}
-    """)
+    """).repartition(1)
     val inputDataset = MetadataUtils.setMetadata(payloadDataset, Extract.toStructType(cols.right.getOrElse(Nil)))
     inputDataset.createOrReplaceTempView(inputView)
 
@@ -163,7 +169,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
         outputView=outputView,
         params=Map.empty,
         persist=false,
-        inputField="inputField"
+        inputField="inputField",
+        batchSize=2,
+        delimiter=delimiter
       )
     ).get
 
@@ -179,7 +187,7 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
     implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
 
-    val dataset = TestDataUtils.getKnownDataset.toJSON.toDF
+    val dataset = TestDataUtils.getKnownDataset.toJSON.toDF.repartition(1)
     dataset.createOrReplaceTempView(inputView)
 
     val actual = transform.HTTPTransform.transform(
@@ -193,7 +201,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
         outputView=outputView,
         params=Map.empty,
         persist=false,
-        inputField="value"
+        inputField="value",
+        batchSize=1,
+        delimiter=delimiter
       )
     ).get
 
@@ -206,7 +216,7 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
     implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
 
-    val dataset = TestDataUtils.getKnownDataset.toJSON.toDF
+    val dataset = TestDataUtils.getKnownDataset.toJSON.toDF.repartition(1)
     dataset.createOrReplaceTempView(inputView)
 
     val thrown = intercept[Exception with DetailException] {
@@ -221,12 +231,14 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
           outputView=outputView,
           params=Map.empty,
           persist=false,
-          inputField="value"
+          inputField="value",
+          batchSize=1,
+          delimiter=delimiter
         )
-      ).get
+      ).get.count
     }
 
-    assert(thrown.getMessage == "HTTPTransform expects all response StatusCode(s) in [200, 201, 202] but server responded with [2 reponses 404 (Not Found)].")    
+    assert(thrown.getMessage.contains("HTTPTransform expects all response StatusCode(s) in [200, 201, 202] but server responded with 404 (Not Found)"))
   }
 
   test("HTTPTransform: validStatusCodes") {
@@ -249,12 +261,14 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
           outputView=outputView,
           params=Map.empty,
           persist=false,
-          inputField="value"
+          inputField="value",
+          batchSize=1,
+          delimiter=delimiter
         )
-      ).get
+      ).get.count
     }
 
-    assert(thrown.getMessage == "HTTPTransform expects all response StatusCode(s) in [201] but server responded with [2 reponses 200 (OK)].")    
+    assert(thrown.getMessage.contains("HTTPTransform expects all response StatusCode(s) in [201] but server responded with 200 (OK)."))
   }     
 
   test("HTTPTransform: Can echo post data - binary") {
@@ -269,7 +283,7 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
     dataset.createOrReplaceTempView(inputView)
     var payloadDataset = spark.sql(s"""
       SELECT *, BINARY(stringDatum) AS value FROM ${inputView}
-    """)
+    """).repartition(1)
     val inputDataset = MetadataUtils.setMetadata(payloadDataset, Extract.toStructType(cols.right.getOrElse(Nil)))
     inputDataset.createOrReplaceTempView(inputView)
 
@@ -284,9 +298,13 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
         outputView=outputView,
         params=Map.empty,
         persist=false,
-        inputField="value"
+        inputField="value",
+        batchSize=2,
+        delimiter=delimiter
       )
     ).get
+
+    transformDataset.cache.count
 
     val expected = transformDataset.select(col("stringDatum")).withColumnRenamed("stringDatum", "value")
     val actual = transformDataset.select(col("body")).withColumnRenamed("body", "value")
@@ -322,7 +340,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
           outputView=outputView,
           params=Map.empty,
           persist=false,
-          inputField="value"
+          inputField="value",
+          batchSize=1,
+          delimiter=delimiter
         )
       ).get
     }
@@ -357,7 +377,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
           outputView=outputView,
           params=Map.empty,
           persist=false,
-          inputField="value"
+          inputField="value",
+          batchSize=1,
+          delimiter=delimiter
         )
       ).get
     }
@@ -394,7 +416,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
           outputView=outputView,
           params=Map.empty,
           persist=false,
-          inputField=inputField
+          inputField=inputField,
+          batchSize=1,
+          delimiter=delimiter
         )
       ).get
     }
@@ -433,7 +457,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
         outputView=outputView,
         params=Map.empty,
         persist=false,
-        inputField="value"
+        inputField="value",
+        batchSize=1,
+        delimiter=delimiter
       )
     ).get
 
@@ -447,66 +473,9 @@ class HTTPTransformSuite extends FunSuite with BeforeAndAfter {
 
     try {
       Thread.sleep(2000)
-      assert(df.first.getString(2).contains("""{"value":0}"""))
+      assert(df.first.getString(1).contains("""{"value":0}"""))
     } finally {
       writeStream.stop
     }
-  }     
-
-  // cannot catch exception yet?
-  //
-  // test("HTTPTransform: Structured Streaming Fail") {
-  //   val thrown0 = intercept[Exception] {
-  //     implicit val spark = session
-  //     import spark.implicits._
-  //     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
-  //     implicit val arcContext = ARCContext(isStreaming=true)      
-
-  //     val readStream = spark
-  //       .readStream
-  //       .format("rate")
-  //       .option("rowsPerSecond", "1")
-  //       .load
-
-  //     readStream.createOrReplaceTempView("readstream")
-
-  //     val input = spark.sql(s"""
-  //     SELECT TO_JSON(NAMED_STRUCT('value', value)) AS value 
-  //     FROM readstream
-  //     """)
-
-  //     input.createOrReplaceTempView(inputView)
-
-  //     val transformDataset = transform.HTTPTransform.transform(
-  //       HTTPTransform(
-  //         name=outputView,
-  //         uri=new URI(s"${uri}/${echo}/"),
-  //         headers=Map.empty,
-  //         validStatusCodes=Option(List(201)),
-  //         inputView=inputView,
-  //         outputView=outputView,
-  //         params=Map.empty,
-  //         persist=false
-  //       )
-  //     ).get
-
-  //     val writeStream = transformDataset
-  //       .writeStream
-  //       .queryName("transformed") 
-  //       .format("memory")
-  //       .start
-
-  //     val df = spark.table("transformed")
-
-  //     try {
-  //       Thread.sleep(2000)
-  //     } catch {
-  //       case e: Exception => println("Here")
-  //     } finally {
-  //       writeStream.stop
-  //     }
-  //   }
-  //   assert(thrown0.getMessage === "HTTPTransform requires a field named 'value' of type 'string' or 'binary'. 'value' is of type: 'date'.")
-
-  // }    
+  }      
 }