+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - added `logger` object to `DynamicConfiguration` plugins.
 - added `customDelimiter` attribute to `DelimitedExtract` and `DelimitedLoad` to be used in conjunction with `delimiter` equal to `Custom`.
 - added optional `description` attribute to all stages.
+- added `batchSize` and `delimiter` to `HTTPTransform` to allow batching to reduce cost of HTTP overhead.
+- refactored `HTTPExtract`, `HTTPTransform` and `HTTPLoad` to fail fast and stay more 'pure'.
 - bump to Alpine 3.9 in `Dockerfile`.
 
 # 1.5.0